% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clean_df}
\alias{clean_df}
\title{Clean a data.frame}
\usage{
clean_df(data, index = NULL)
}
\arguments{
\item{data}{data.frame to sort}

\item{index}{sorting columns (character vector)}
}
\description{
Sorts rows, cleans columnames, alphabetizes columns, remove empty rows and
columns of a data frame
}
\examples{
sort_df(dat, c('iso3c_host', 'iso3c_home'))
}
\keyword{data}
\keyword{frame}
\keyword{sort}
