% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ClosestString}
\alias{ClosestString}
\title{Find best fuzzy matches for each element of a vector with itself}
\usage{
ClosestString(x, y, n = 1)
}
\arguments{
\item{x}{strings to be matched (character)}

\item{y}{strings to match to (character)}

\item{matches}{how many strings to return? (integer)}
}
\description{
Find best fuzzy matches for each element of a vector with itself
}

