% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{confint_robust}
\alias{confint_robust}
\title{HC-robust confidence interval}
\usage{
confint_robust(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{fitted model compatible with the sandwich package}
}
\description{
Computes confidence intervals using the sandwich package
}

