% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\name{get_dups}
\alias{get_dups}
\title{Convert percentage to letter grade}
\usage{
get_dups(data, index)
}
\arguments{
\item{data}{(data.frame)}

\item{index}{(vector of columns names)}
}
\description{
Convert percentage to letter grade
}
