% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{GetWDI}
\alias{GetWDI}
\title{Download and clean WDI data}
\usage{
GetWDI(dictionary = c(NY.GDP.PCAP.KD = "gdppc"))
}
\arguments{
\item{dictionary}{(named character vector)}
}
\description{
Download and clean WDI data
}
\examples{
GetWDI(dictionary=c("NY.GDP.MKTP.CD.XD"='deflator', "NY.GDP.PCAP.KD"='gdppc'))
}

