% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{ggplot_add.conditional_theme_vab}
\alias{ggplot_add.conditional_theme_vab}
\title{Conditional formatting of axes for facets}
\usage{
\method{ggplot_add}{conditional_theme_vab}(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add \code{object} to}

\item{object_name}{The name of the object to add}
}
\description{
Conditional formatting of axes for facets
}
\keyword{internal}
