% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\name{NormalizeGrades}
\alias{NormalizeGrades}
\title{Normalize grades by 2nd/3rd quartiles}
\usage{
NormalizeGrades(data, plot = FALSE)
}
\arguments{
\item{data}{(data.frame with 3 columns: 'id', 'note', 'correcteur')}

\item{plot}{(boolean plot or return raw data)}
}
\description{
Normalize grades by 2nd/3rd quartiles
}

