% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{pct2let}
\alias{pct2let}
\title{Convert percentage to letter grade}
\usage{
pct2let(grades, slack_fail = 2, slack_pass = 1)
}
\arguments{
\item{grades}{(numeric)}

\item{slack_fail}{(integer)}

\item{slack_pass}{(integer)}
}
\description{
Convert percentage to letter grade
}
