% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_last_csv}
\alias{read_last_csv}
\title{Read last CSV}
\usage{
read_last_csv(file, ...)
}
\arguments{
\item{file}{beginning of file name -- no extension (character)}
}
\description{
Globs file names, sorts and reads the last file. Useful when CSV file names
have dates appended to them.
}
\examples{
read_last_csv('dataset')
}
