% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_last_tsv}
\alias{read_last_tsv}
\title{Read last CSV}
\usage{
read_last_tsv(path)
}
\arguments{
\item{path}{stem to read from (without date or .tsv extension)}
}
\description{
Globs file names, sorts and reads the last file. Useful when CSV file names
have dates appended to them.
}
\examples{
read_last_tsv('dataset')
}
