% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sort_df}
\alias{sort_df}
\title{Sorting a data.frame based on an index}
\usage{
sort_df(data, index = NULL)
}
\arguments{
\item{data}{data.frame to sort}

\item{index}{sorting columns (character vector)}
}
\description{
This function sorts the columns and rows of a data.frame using an index
}
\examples{
sort_df(dat, c('iso3c_host', 'iso3c_home'))
}
\keyword{data}
\keyword{frame}
\keyword{sort}

