% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{tables}
\alias{tables}
\title{Saves (multiple) LaTeX tables to file}
\usage{
tables(models, filename = NULL, label = "", caption = "",
  output = "latex", varcov = NULL, nmod = 6, digits = 3, stars = NULL,
  custom.coef.map = NULL, include.loglik = TRUE, include.aic = FALSE,
  include.bic = FALSE, include.deviance = FALSE, use.packages = FALSE,
  warn = TRUE, ...)
}
\arguments{
\item{output}{txt, html, or tex}

\item{varcov}{a list of variance covariance matrices of length length(models)}

\item{nmod}{maximum number of models (columns) per table}
}
\description{
Saves (multiple) LaTeX tables to file
}
