% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{TeXTables}
\alias{TeXTables}
\title{Saves (multiple) LaTeX tables to file}
\usage{
TeXTables(models, file, label, caption, varcov = NULL, dict = NULL,
  nmod = 6, digits = 3, stars = NULL,
  omit.coef = "iso3|region|president|year", include.loglik = FALSE,
  include.aic = FALSE, include.bic = FALSE, include.deviance = FALSE,
  use.packages = FALSE, ...)
}
\arguments{
\item{varcov}{a list of variance covariance matrices of length length(models)}

\item{nmod}{maximum number of models (columns) per table}
}
\description{
Saves (multiple) LaTeX tables to file
}

