% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{theming}
\alias{theming}
\title{Set themes for base R graphics and ggplot2}
\usage{
theming(palette = "batlow", begin = 0, end = 1, direction = -1)
}
\arguments{
\item{palette}{scico color scheme name (character)}

\item{begin}{numeric [0, 1]}

\item{end}{numeric [0, 1]}

\item{direction}{-1 or 1}
}
\description{
Set themes for base R graphics and ggplot2
}
