% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_csv}
\alias{write_csv}
\title{Write to CSV file with automatic date}
\usage{
write_csv(data, file, row.names = FALSE)
}
\arguments{
\item{data}{data.frame to write}

\item{file}{filename to write to}

\item{include}{row.names (boolean)}
}
\description{
Writes to a CSV file with date automatically appended to file name
}
\examples{
write_csv(dat, file="data/dataset")
}

