% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_df}
\alias{write_df}
\title{Write data frame to CSV file with date stamp}
\usage{
write_df(data, file, row.names = FALSE)
}
\arguments{
\item{data}{data.frame to write}

\item{file}{filename to write to}

\item{include}{row.names (boolean)}
}
\description{
Writes to a CSV file with date automatically appended to file name
}
\examples{
write_df(dat, file="data/dataset")
}

