% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_tsv}
\alias{write_tsv}
\title{Write data frame to TSV file with date stamp}
\usage{
write_tsv(data, path)
}
\arguments{
\item{data}{data.frame to write}

\item{path}{stem to write to (without date or .tsv extension)}
}
\description{
Writes to a TSV file with date automatically appended to file name
}
