% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\name{pct2let}
\alias{pct2let}
\title{Convert percentage to letter grade}
\usage{
pct2let(grades, slack = 1.5)
}
\arguments{
\item{slack}{(integer)}

\item{index}{(character vector)}
}
\description{
Convert percentage to letter grade
}
